Use StringOp::Builder class to reduce executable size
Use new StringOp::Builder class to construct a string from several pieces. Before
we used combinations of string::operator+() and StringOp::toString().

The main advantage of StringOp::Builder is that we made all methods non-inline
(all uses are in non-performance critical paths). This saves a couple of bytes
per call, but since there are so many of them, it saves in total about 30kB.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10392 7512d703-071c-0410-b5c2-d2bdd895b7d1