Fixed bug [2844043] Firebird small graphics corruption
The problem is caused by the temporal decoupling between CPU and cmdEngine,
more specifically by a missing sync between these two subsystems.

It can be triggered in the following scenario:
 - VRAM location 'x' contains value 'v1'
 - VDP is executing a command that will overwrite location 'x' with value 'v2'
   (v1 != v2). This write will happen at time 't1'.
 - At time 't2' (t1 < t2) the CPU also writes to VRAM location 'x'. It writes
   the original value 'v1' again.
Now cmdEngine is still somewhere before 't1' in time and CPU is already at
't2'. As an optimization the cpuWrite() routine checks whether the VRAM already
contains the about to be written value, if so it does nothing (also NOT syncing
with cmdEngine, this is the bug!). Some time later cmdEngine is synced, and it
writes the wrong value (an older value) to VRAM.

Fixed by moving the 'newValue == oldValue' check after the cmdEngine sync. This
allowed to move the check from cmdWrite() and cpuWrite() to a single place in
the writeCommon() routine.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10438 7512d703-071c-0410-b5c2-d2bdd895b7d1