Oops, some LZO_ABI_* and LZO_OS_* macros were still used.
They were removed in revisions 10545 and 10548. These were used to define the
lzo_uint type. However I verified that in the (remaining) code simply using
'unsigned' for all platforms should work just fine (at least if the memory
block is less than 4GB in size). This is also more in line with the rest of the
openMSX code.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10558 7512d703-071c-0410-b5c2-d2bdd895b7d1