Removed types from lzoconf.
The typedefs (or macros) in lzoconf.hh didn't really make the code any clearer,
so I directly use the underlying type in the rest of the code. lzoconf.hh file
is now empty so I removed it.

Also did a compilation fix for vc++: it doesn't make sense to add a __restrict
qualifier on non-pointer parameters. vc++ also doesn't like this, it gives
compilation errors on it.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10560 7512d703-071c-0410-b5c2-d2bdd895b7d1