Fixed UMR.
This could even lead to a crash. I could easily trigger it by starting
  openmsx VKILLER.ROM.gz -machine msx2
then in the console execute 'revere start' and let it run for some minutes (or
use F9).

I believe this bug was already there in the original code. Maybe now it's
easier to trigger because the 'workbuffer' has moved: before it was allocated
by the caller, now it's allocated inside the compression function.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10563 7512d703-071c-0410-b5c2-d2bdd895b7d1