replace 'assert(false)' with 'UNREACHABLE'
Gcc-4.5 introduced support for __builtin_unreachable(). This statement indicates
that the program control flow cannot reach this statement and the compiler is
allowed to make optimizations based on this assumption.

We defined a macro 'UNREACHABLE' that expands to either __builtin_unreachable()
or 'assert(false)'. It expands to __builtin_unreachable() only when that
statement is available (gcc-4.5) and when asserts are disabled (IOW when the
assert marco itself would expand to nothing). So in debug/devel flavours we
still trigger these asserts.

A typical use-case is:
  switch (<some-condition>) {
    case <a>: ...
    case <b>: ...
    ...
    default: UNREACHABLE;
  }

Compared to using 'assert(false)' this results in both faster and smaller code.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10580 7512d703-071c-0410-b5c2-d2bdd895b7d1