Added romblock debuggable
This is 64kB sized debuggable with as name '<romname> romblocks'. Every byte
in this debuggable shows the selected rom block for this rom mapper. So for
a mapper with 16kb pages, all values in the same 16kb region will have the
same value.

Some mappers also allow to map SRAM (or even some other stuff). What this
debuggable will return depends on the actual mapper type, but mostly it
returns the value that was written to select the SRAM (often a higher value
than for rom blocks).

Some regions of the mapper may be unmapped (often region 0x0000-0x3fff and
0xc000-0xffff). This debuggable will return 0xff for such regions.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10587 7512d703-071c-0410-b5c2-d2bdd895b7d1