Fixed 'romblocks' debuggable for the 'Game Master 2' mapper type.
Game Master 2 is a 8kB block size mapper. Though it has only 4kB SRAM (which
is mirrored twice in one 8kB block). To be able to implement this with the
generic RomBlocks base class, it is internally implemented as a 4kB mapper.

The only problem with this approach is that the romblocks debuggable was
wrong. This is fixed in this patch: it's now possible to indicate how
much bigger the logical block size is compared to the implementation size.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10593 7512d703-071c-0410-b5c2-d2bdd895b7d1