Fix superimpose in screen 0.
Rationale:

The difference between the palFg and palBg is that index 0 of palFg is 
the background color, so after "color ,4,4" then palFg[0] will be the 
value of palBg[4], unless transparency of color 0 has been disabled in 
bit 5 of r#9, in which case it will be palBg[0].

If the The color of the background in Text1 is != 0, then nothing has
changed. If the background is 0, then necessarily palFg[0] is the same as 
palBg[0].

Note that superimposing on the MSX2 has not been taken into consideration. 
Most likely if r#9 = 0x20, then color 0 is not transparent, thus not 
superimposing, although this code would. This needs testing when MSX2
superimposing is implemented.


git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10595 7512d703-071c-0410-b5c2-d2bdd895b7d1