Make 'after' command more 'Tcl compatible'
Tcl has an 'after' command. In openMSX we override this with our own 'after'
command. The two are very similar, but not 100% compatible. This patch makes
them more (but not yet fully) compatible. We now also accept:
  after <ms> <script> ...

Looking back it was probably a mistake to override the Tcl after command.
Though it's like this for historical reasons: before we integrated the Tcl
interpreter in openMSX, we already had an after command with this syntax,
which we kept for backwards compatibility.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10605 7512d703-071c-0410-b5c2-d2bdd895b7d1