Removed StringPool
StringPool is a mechanism to only store one actual copy in memory for a whole
group of otherwise identical strings. For example in our rom database we have
lots of identical strings. Also in our XML data structures (class XMLElement) we
have lots of identical tag names.

On my machine (Linux x86_64) using StringPool saved about 800kB memory, but it
also slows down startup time from 0.24s to 0.34s. For PCs these numbers don't
matter much, they are more important for small systems like Dingoo. For Dingoo
we don't have very accurate numbers, but very roughly: StringPool saves ~600kB
but slows down startup from 8s to 9s.

It's not clear which trade-off is best. Though in the future we plan to use
sqlite for the romdatabase instead of parsing the XML and keeping it in memory.
So the advantage of StringPool will disappear. That's why we decided to already
remove StringPool now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10617 7512d703-071c-0410-b5c2-d2bdd895b7d1