Use compile-time constant color key
Before this patch we used an otherwise unused color value as the color key
(color key = special value meaning transparent). There was also some special
logic to avoid collisions between the color key and the actual colors while
(pre)calculating the MSX palette, in case of a collisions, a new color key
was chosen.

This patch makes the color key a compile-time constant. Collisions between
this color key and the palette colors are now resolved by slightly changing
the palette.

In 16bpp we choose '1' as color key and in case of a collision the palette
color is replaced with '0'. Visually colors '0' and '1' are very close, almost
indistinguishable (this is true in practically all color encodings (encodings
like RGB555, RGB565, ...)).

In 32bpp there cannot be collisions between the color key and the palette. All
palette colors have a fully opaque alpha channel (all alpha bits set). As color
key we choose 0 (so alpha bits are zero). In the future I'd like to use a true
alpha channel to indicate transparency instead of a special color key value.
Choosing a color key with alpha=0 is already a small step in that direction.

The main reasons for choosing a compile-time constant color key are:
 - Simplicity. The collision avoidance code is much simpler.
 - Speed. While rendering frames, we often need to compare against the color
   key. Depending on the CPU, comparisons with a constant are cheaper than
   with a variable. Especially with a 'simple' constant like '0' or '1'.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10627 7512d703-071c-0410-b5c2-d2bdd895b7d1