For YM2413/Y8950 output melody and rhythm on separate channels
YM2413 and Y8950 can work in two modes:
 - 9 melody channels, no rhythm channels
 - 6 melody channels, 5  rhythm channels

Before this patch, we assigned the melody channels to the openmsx output
channels 1-9 and the rhythm channels to 7-11. Note that there's an overlap
of the channels 7-9, but that wasn't a real problem because these anyway
couldn't be used simultaneously.

In this patch we assign melody to 1-9 and rhythm to 10-14. So we created
3 extra channels (of which at least 3 are always idle).

The reason for doing this is the FM-stereo-PAK: it outputs the melody on one
side and the rhythm on the other side. So we need separate output channels to
be able to assign different balance settings to them. FYI: the real YM2413 chip
has two output pins: one for melody and one for rhythm.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10642 7512d703-071c-0410-b5c2-d2bdd895b7d1