YM2413Okazaki refactoring
Don't store 'type' as member of the Slot class. Instead calculate it when it's
needed. This is also how the YM2413Burczynski core handles it. Also renamed it
to 'actAsCarrier'.

TODO this refactoring might have a functional change when leaving rhythm mode
while 'slot_on' is still active. Check if this is really the case and
investigate what happens on the real HW.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10652 7512d703-071c-0410-b5c2-d2bdd895b7d1