YM2413Burczynski refactoring
Moved calculation of envelope/phase into sound generation (was in a seperate
advance() method). This is also how the YM2413Okazaki core does it. It also possibly
generates faster code (not verified).

TODO verify if this caused any functional changes. If it did the difference
will be very small and it's not clear (and difficult to verify) whether this is
closer to the real HW or not.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10655 7512d703-071c-0410-b5c2-d2bdd895b7d1