YM2413Okazaki refactoring (EXPERIMENTAL)
Largly speedup calc_envelope() by making specialized versions for the case when
the envelope generator is in a 'fixed state' (the value of the envelope won't
change).

These 'fixed envelope' routines are much simpler than the general ones. And,
depending on the song of course, about half of the time it's possible to use
the 'fixed envelope' versions. So this results in very measurable speedup (a
quick test showed about 10% speedup on total emulation time, with
renderer=none). OTOH the resulting binary is about 90kB larger.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10658 7512d703-071c-0410-b5c2-d2bdd895b7d1