Removed dphaseTable from YM2413Okazaki core.
We don't need a 256kB lookup table for a relatively simple calculation. This
calculation is performed often (on most register writes), but not during sound
generation.

In the past, when sound generation didn't run yet on the native chip frequency,
this calculation was much more expensive (required floating point operations).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10665 7512d703-071c-0410-b5c2-d2bdd895b7d1