Copied lfo_am implementation from YM2413/YMF262 to Y8950
All OPLx cores now use the same lfo_am implementation. This new implementation
was verified on a real YM3812 (OPL2), but it's very likely to be the same on
YM2413/YMF262/Y8950 because it very closely matches the parameters for lfo_am
in the datasheets. Also this implementation is much more suited for a HW
implementation than the old one.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10667 7512d703-071c-0410-b5c2-d2bdd895b7d1