Fixed possible crash on loadstate with RomPanasonic mapper
Could be reproduced by:
  start Panasonic_FS-A1GT MSX machine
  boot into the built-in MSX-VIEW firmware
  do savestate, loadstate

Revision 8219 implemented a mechanism to not have to rely on the order of
instantiation of MSXDevices, otherwise the PanasonicRam device should be
instantiated before the RomPanasonic device. This worked for the savestate
case, but not for loadstate (and load only fails when the RAM is actually
currently mapped into the rom mapper, so initial tests didn't show it). BTW
revision 8219 did not introduce this bug, it just didn't solve the existing
problem correctly.

Fixed now by simply making the instantiation order mandatory. This order was
already correct in the hardwareconfig.xml files. But I've also added an
explicit dependency in those config files, so that this order stays correct,
even after future changes of those config files.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10739 7512d703-071c-0410-b5c2-d2bdd895b7d1