Removed CLI event recorder/replayer
The previous commit changed the RecordedCommand event handling, so the
recorder/replayer is now broken. This feature was also not that useful because
it was only possible to use it from the command line. And it also always
started from a MSX boot.

I removed this feature for now. We should replace it with a record/replay
mechanism starting from a savestate (possibly coupled to the reverse
mechanism).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10748 7512d703-071c-0410-b5c2-d2bdd895b7d1