Fixed 'hanging key' problem when end of replay log is reached
Similar to previous commit: if the replay log ends with a pressed MSX key,
there won't be a corresponding key-up event and the MSX key stays pressed.

Solution is to also resync with host keyboard (and other input devices) when
the end of the replay log is reached. (Implemented by ending the replay log
with a special end-of-log event).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10752 7512d703-071c-0410-b5c2-d2bdd895b7d1