Fixed off-by-one error in ReverseManager
The 'collectCount' variable used to indicate 'the index of the NEXT to be taken
snapshot'. When going back in time, we set this variable to the number of taken
snapshots. So when the next snapshot is taken it will overwrite the current one.

The code is is easier to understand if the variable simply means 'number of
taken snapshots'. Thus fixed by changing the meaning of the variable (and
changing the rest of the code accordingly).

The code was written like that to be able to distinguish the states
 - we're not collecting (reverse start command was not given)
 - we're collecting but don't have the first snapshot yet
Though this second state only exists very briefly and never when
entering/leaving a public method. So it's not really required to have separate
states for these two cases.

(This does not solve the 'long replay not always identical' problem.)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10774 7512d703-071c-0410-b5c2-d2bdd895b7d1