Fixed possible crash on loadstate
The Renderer class is not part of a savestate (we can't easily include it
because the type of renderer can be different on each loadstate). On loadstate
it's possible (even likely) that the VDP is somewhere in the middle of a frame.
To prevent 'bad things' from happening we suppress rendering till the beginning
of the next frame.

That's the basic idea anyway. It worked well expect for a call to
Renderer::reset() from VDP::serialize which also triggered a
Renderer::frameStart(). Fixed by renaming Renderer::reset() to
renderer::reInit() and removing the call to frameStart(). See comments in code
for more details.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10778 7512d703-071c-0410-b5c2-d2bdd895b7d1