Preparation to implement superimpose for SDLGL-PP
Currently superimpose is only implemented for the SDL renderer and with a
special 'transparent scaler' (meaning that you don't get any of the effects of
the user-selected scaler). It would be nice if superimpose also worked with the
SDLGL-PP renderer and that you still get the visual effects of your selected
scaler algorithm.

It's not immediately clear how superimpose should be combined with an arbitrary
scaler algorithm. E.g. should superimpose come before or after (MSX-frame)
scaling in the rendering pipeline.
 * Before: MSX frame and external frame are first combined and then scaled.
   Problems:
   - MSX and external frame may not have the same resolution, before passing
     line(s) to the scaler, they must be combined to line(s) of the same
     resolution. We can choose the MSX resolution, the external frame's
     resolution or yet some other resolution. But each choice will give visual
     artifacts (for example when switching superimpose on/off).
   - effects like 'hq', 'hqlite', 'SaI', 'ScaleNx' may not look very good on
     typical external video frames.
 * After: MSX frame is first scaled and then combined with the external frame.
   Problems:
   - Effects like 'simple', 'tv', 'rgb' will look very weird when only applied
     to the MSX frame but not to the external frame.

To solve this we let each scaler algorithm decide for itself how to combine
superimpose and MSX-scaling. For example the 'simple' scaler can apply blur
(only) to the MSX image but the scanline effect to the combination of MSX/ext
frame. The 'hq' scaler can first scale the MSX frame and then combine it with
the external frame.

As an experiment I've implemented these scaler specific combinations as openGL
shaders (so for the SDLGL-PP renderer). It turns out this works very well. I
will commit these SDLGL-PP shaders in follow-up patches. Though it remains to be
seen how easy it is to implement the same algorithms in software (so for the
SDL renderer).

This patch is only a preparation to make scaler-specific superimpose possible.
It changes the code so that instead of only telling the VDP that superimpose
is enabled/disabled, it also tells with which frame the MSX frame should be
combined. The VDP can then pass this information (indirectly) to the scaling
algorithm (VDP->renderer->rasterizer->postprocessor->scaler). In this patch
that information is not yet used, it will be in follow-up patches.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10786 7512d703-071c-0410-b5c2-d2bdd895b7d1