Support superimpose in SDLGL-PP 'simple' scaler
See commit message of previous commit for much more details.

This teaches the 'simple' scaler about superimpose. This scaler has two main
parts: blur and scanlines. The blur part is only applied to the MSX image, then
the blurred MSX image and the original superimpose frame are combined and
finally scanlines are applied.

Because this is the first scaler that's converted, this patch touches a lot of
files. Only a small part of these are actual 'simple scaler' changes. Another
part is adding extra helper functions that will also be useful for other
scalers. And another large part is because the GLScaler::scaleImage() method
got extra parameters. These new parameters are not yet used by the other
scalers, but the files had to be changed nevertheless.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10787 7512d703-071c-0410-b5c2-d2bdd895b7d1