Made noise calculation correct for 32bpp
The noise calculation ('noise' setting, not the snow effect) for the SDL
renderer in case of no MMX nor SSE (so the pure C++ version) did only calculate
the lower 24bits (in case of 32bpp mode). On some systems the highest 8 bits do
contain a color component (instead of the alpha channel). For example in OSX
64bit, bits 24-31 contain the blue color component. So before this patch, noise
went wrong on those systems.

The fix consists out of two main parts:
 - In the pre calculated noise table, make sure the alpha component always has
   value zero and only the color components contain actual noise values.
 - When applying the noise to the image, apply it to all 32 bits.

The latter part was quite tricky. It was not easily possible to extend the old
code to also work on the upper 8 bits (because of overflow). The new approach
contains some clever bit-twiddling-tricks to avoid this. As a bonus it's also
faster than the old code. See comments in the new addNoise4() helper function,
to learn about these tricks.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10820 7512d703-071c-0410-b5c2-d2bdd895b7d1