Reset input device state on a 'real' loadstate
When a savestate was created with some keys pressed, and later loaded again
the MSX keys will still be pressed, even though the corresponsing host keys
are no longer pressed (similar for other input devices).

For reverse/replay this is the required behavior. But for a normal loadstate
it's better to let the MSX resync with the host input state.

Actually before this patch, we did already resync. But only on the first
input event. Now we resync immediately after loadstate.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10831 7512d703-071c-0410-b5c2-d2bdd895b7d1