Fixed assert when using reverse in "Aleste Special"
This assert could relatively easily be triggered in "Aleste Special", while
playing the game (not the intro demo). Seems to be related to savestate with an
in-progress VDP command.

The underlying problem was that de-serializing VDPVRAM indirectly triggered a
VDPVRAM::sync() which in turn indirectly triggered a
SpriteChecker::checkUntil(), but SpriteChecker was not yet de-serialized. Fixed
by swapping the order of de-serialization of VDPVRAM and SpriteChecker.

A second problem I found while debugging this problem, was a problem that only
occurred in debug builds: VDPVRAM has an extra 'vramTime' member (only) in
debug builds. This member is not (de)serialized. In the constructor it is
initialized with the 'current' time. But because of the temporal decoupling of
the different components there is no well defined current time. And in case of
a snapshot with a VDP command in progress, 'vramTime' should not get the
'current' time (= the time of the Scheduler), but the time of VDPCmdEngine
(which is earlier). Fixed by simply initializing 'vramTime' to zero.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10847 7512d703-071c-0410-b5c2-d2bdd895b7d1