Use event mechanism to cleanup old MSXMotherBoards
In revision 7897 we introduced a mechanism to clean up old MSXMotherBoards in a
safe way.  See that commit for details, but in short: we only clean up when
there are no pending events.

But in case there are really a LOT of events, or when one event (possibly)
triggers queueing of another event, there can be a lot of time between
cleanups. In this patch we move the cleanup itself to an event handler, so it
gets done in between the handling of the other events.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10850 7512d703-071c-0410-b5c2-d2bdd895b7d1