Use dynamic instead of static polymorphism in many SDL scalers
See documentation of the PolyLineScaler class (src/video/LineScalers.hh) for
more details. In short:
 - static polymorphism (c++ templates) often results in faster but bigger code
 - dynamic polymorphism (c++ virtual functions) often results in slower but
   smaller code.
For this specific case I couldn't really measure a difference in speed, while
the code was about 70kB bigger. So I changed the code to use dynamic
polymorphism.

The main motivation for this change is the upcoming work on superimpose with
SDL scalers. Using dynamic polymorphism there would minimize code bloat and
judging from this patch, the speed impact should be very small.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10941 7512d703-071c-0410-b5c2-d2bdd895b7d1