Fixed assert in turbor 32bit debug flavour
In a 32-bit executable build with debug flavour, just starting a MSX turboR
machine triggered an assert after some time (before the BASIC prompt was
reached).

The reason was that after running Z80 for some significant amount of time and
then switching to R800 mode, the [lastRefreshTime .. currentTime] interval was
so large that some calculation didn't fit in a 32-bit value. We have an
optimized routine to perform 64-bit by 32-bit divisions (only on 32-bit, on
64-bit machines this optimization is not useful), though this only works when
the result of that division fits in a 32-bit value. (Only in a debug flavour)
there is an assert that checks whether this assumption was indeed true, so this
assert was triggered in the above scenario.

Fixed by synchronizing the 'lastRefreshTime' variable after a Z80->R800 switch,
so that the normal code path never sees such huge time intervals anymore.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10959 7512d703-071c-0410-b5c2-d2bdd895b7d1