Fixed "No setting xxx on current machine" error
Could be triggered like this:
  - start openmsx with e.g. cbios machine
  - in openMSX console:
      > set renshaturbo 20    current machine doesn't have this setting,
                              so, for now, this is a normal Tcl variable
      > create_machine
      > machine2::load_machine Panasonic_FS-A1GT
      Error in "Panasonic_FS-A1GT" machine: No setting 'renshaturbo' on current machine

Error is fixed now. Though in the above scenario the 'renshaturbo' setting will
have the value '0' (the default value) instead of the value '20'. This may seem
strange with it is consistent with the following scenario.

  - start openmsx (any machine will do)
  - in openMSX console:
      > set PSG_volume 100
      > create_machine
      > machine2::load_machine cbios
      > set machine1::PSG_volume
      100                               the value we just set
      > set machine1::PSG_volume
      75                                the default value

So in both scenarios after loading a machine, the per-machine settings all have
their default value (whether that setting already existed in the previous
machine or not).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11045 7512d703-071c-0410-b5c2-d2bdd895b7d1