Fixed possible out-of-array-bounds write
The method FrameSource::getlinePtr(line, width) can return a line from a
buffer scaled to the requested width. If required this method internally
allocates extra memory buffers.

The problem was that the maximal width of such a buffer was set to the
maximal line with stored in the FrameSource. But when you for example
request to up-scale a V99x8 line (max width = 640) to 960 (for scale
factor 3), we write past the end of the buffer.

Fixed by always allocating enough space to hold a line of 1280 pixels.
This is large enough for all possible scalings.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11091 7512d703-071c-0410-b5c2-d2bdd895b7d1