Teach SDL-renderer-scalers about super-impose.
For now super-impose is implemented in the base classes of the scalers (for 1x,
2x and 3x) as follows:
 - normally scale the MSX frame,
 - then perform (per-pixel) alpha-blending with the video frame.
This approach is fine for the scale algorithms SaI, scaleNx, hq and hqlite. But
doesn't give the best visual result for simple and RGB. I'll change the latter
two in the next commits. (There is no implementation for the TV scaler for SDL).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11104 7512d703-071c-0410-b5c2-d2bdd895b7d1