Implemented visually better super-impose for SDL RGB and simple scaler
Instead of performing the alpha-blending after scaling of the MSX frame, we
first blend the MSX frame with the video frame and only then apply the
scale algorithm. So now the scanlines will also be drawn over the video
layer and the for the RGB scaler, the RGB effect is also applied to the
video layer.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11107 7512d703-071c-0410-b5c2-d2bdd895b7d1