Removed TransparentNxScaler scalers
In the past superimposing (transparent pixels in the MSX image) were handled by
special scalers that simply don't draw the transparent MSX pixels. So the video
layer had to be drawn first and the MSX image was drawn on top. Now all normal
scalers know how to handle superimpose. So the transparent scalers are no
longer required.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11108 7512d703-071c-0410-b5c2-d2bdd895b7d1