frameStart() related cleanups
In frameStart() instead of calling frameStartTime.advance_fast(), use
frameStartTime.reset(). The latter is faster because it doesn't need to round
the time to a 'clock-tick-boundary'. In this case we are sure that the given
time is at such a boundary.

In the de-serialization code, don't use frameStart(), instead set
frameStartTime directly and only executed to useful side-effects of
frameStart(). This should make the code easier to understand.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11140 7512d703-071c-0410-b5c2-d2bdd895b7d1