Simplified VDPVRAM::serialize()
The old code was probably not wrong, but it was very subtle: in
VRAMWindow::serialize(), the members 'effectiveBaseMask' and 'combiMask' were
calculated based on the wrong 'sizeMask'. Though later it got corrected.
The new code first restores sizeMask so VRAMWindow::serialize() immediately does
the right thing.

It would be even better to not call methods with side-effects at all from the
de-serialization code. That's for a future refactoring.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11143 7512d703-071c-0410-b5c2-d2bdd895b7d1