Fixed problem with loading an old manbow2 savestate
Quibus gave me a savestate that worked in the past (could be loaded
successfully), but now fails with an assert. Here is a description of what went
wrong:

  The savestate had a rom cartridge extension for manbow2. Internally in a
  savestate, extensions are stored as two main parts:
   - the config for the extension (a bit like the hardwareconfig.xml files)
   - the actual state of the devices in the extension
  (Before this patch) the mapper type in the config part for roms was kept at
  the default value 'auto'. And in this particular savestate the mapper was
  detected as 'KonamiSCC', so the devices-state part of the savestate had
  information for a KonamiSCC mapper.

  Some time after this savestate was created, we probably updated the rom
  database. So now when you load the savestate, it auto-detects the mapper as a
  'Manbow2' mapper. But then when it tries to restore the state of that mapper
  it notices that the stored state is actually for a KonamiSCC mapper, and that
  triggers the assert.

I've fixed this by instead of storing the default value 'auto' for the
mapper type in the config parts of a savestate, storing the actual detected
mapper type. This won't fix this particular savestate, but it should prevent
similar problems in the future.

We should also change openMSX to instead of asserting on such savestates, print
some kind of error message. I'll do that in a later patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11144 7512d703-071c-0410-b5c2-d2bdd895b7d1