Don't store rom type in MSXRom anymore.
We used to store a RomType enum value in the MSXRom base class. This was only
used for the 'machine_info device <some-msx-rom-device-name>' command. In a
previous patch we already store the actual type in the device configuration.
So we can as well use that. This simplifies the code a bit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11146 7512d703-071c-0410-b5c2-d2bdd895b7d1