Implemented alpha-fading for whole OSD widget hierarchies.
!! This change is not fully backwards compatible.     !!
!! All Tcl scripts in share/scripts have been updated !!

Before this patch, OSD widgets had a -fadeTarget and -fadePeriod property that
changed the -alpha property (of only that widget) over time. Now there is an
extra property -fadeCurrent and only that property is changed over time (-alpha
is left unchanged).

Also note that -alpha is an integer in the range 0..255, while -fadeCurrent is
a float in the range 0..1.

The 'actual' transparency of a widget depends on its -alpha property and the
-fadeCurrent property of itself and all its hierarchical parents.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11174 7512d703-071c-0410-b5c2-d2bdd895b7d1