Fixed possible assert on replay
We tried to skip (re)creating snapshots during replay. Though because the
EmuTime for the snapshots is not the same during replay (see comment in code)
this is a bit tricky. Though skipping these snapshots is only a minor
optimization, so it's probably easiest to not do this optimization at all
(also makes the code easier to understand).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11175 7512d703-071c-0410-b5c2-d2bdd895b7d1