Try to reuse osd elements in reverse bar
The reverse bar has a variable number of tickmarks. Before this patch, on each
redraw, all the osd elements for the tickmarks were destroyed and new elements
were created. Now we only create/destroy the minimum amount (and reuse the
other). This also fixes a bug on
  toggle_reversebar ; reverse stop ; toggle_reversebar

This patch contains two other tweaks:
 - disable reverse bar on 'reverse stop'
 - fade-in reverse bar a little faster

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11201 7512d703-071c-0410-b5c2-d2bdd895b7d1