Added alternative SSE2 implementation for YCbCr->RGB conversion
Compared to the old code, this one needs less temporary memory stores.

Possibly it has better instruction scheduling, though I'm not 100% sure about
this. I originally wrote this using SSE2 intrinsics, so gcc did the original
instruction scheduling. Then I took the generated code and tweaked it some
more. Mostly to eliminate register spills and to make it fit in only 8
registers (so that it also works in 32 bit mode).

I also removed the prefetch instructions because they didn't seem to help. The
data was anyway already very likely in the CPU cache because the frame was only
just decoded.

Although I think it should be possible to write a MMX version of this code, I
didn't bother. The old code did have a MMX implementation, but that one is
removed now.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11230 7512d703-071c-0410-b5c2-d2bdd895b7d1