Key presses that have a function in the console don't trigger hotkeys anymore.
Before this patch, hotkey events had higher priority than console events. So
keyboard events were first send to the hotkey subsystem and if there was
nothing bound to that event, it was send to the console. Now we swapped the
priority. So we first send the event to the console and only then, if it's not
consumed by the console, we send it to the hotkey subsystem. The 'if not
consumed by the console' stuff is the biggest part of this patch.

The main motivation for this change is to be able to use pgup/down as hotkeys
for reverse_prev/next. Before this patch, when you would bind something to
pgup/down, you also lost the ability to scroll up/down in the console. Another
annoyance that's now solved is that you can revert a binding like "unbind t
exit", otherwise, if you tried to type "unbind t" openmsx would already exit.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11232 7512d703-071c-0410-b5c2-d2bdd895b7d1