Fixed crash in 32-bit builds when EmuTime::dummy() was actually used.
Usually we pass EmuTime::dummy() function parameter to functions that don't
actually use this parameter. But in ReverseManager.cc it was also used to
initialize a EmuTime variable to some dummy value. This works fine in 64-bit
builds, but crashes in a 32-bit build. (Could for example be triggered with
"reverse goback 1").

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11236 7512d703-071c-0410-b5c2-d2bdd895b7d1