Fixed regression in yuv2rgb conversion
There was a signed-overflow in the calculation of the blue channel (only) when
the U input component had the maximum value (255). We calculate (with 16-bit
precision):
   (u * 129) >> 6   with u = [0 .. 255]
Before this patch we used a signed shift (still required for the other color
components). Now we use an unsigned shift.

In the original code we used an alternative formula:
   ((u - 128) * 129) >> 6
This formula doesn't have any problems with signed-overflow.

Thanks to seanie for spotting this regression and bisecting it.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11246 7512d703-071c-0410-b5c2-d2bdd895b7d1