Fixed inconsistency between MSXMusic and MSXFmPac serialization version
This is actually a limitation of the ArchiveBase<T>::serializeInlinedBase()
method in the serialization framework (see documentation in serialize.hh for
more details). This patch adds a (compile-time) check that
serializeInlinedBase() is used with a consitent version number. This method
should only be used rarely.

Bug could be triggered by doing:
	reverse savereplay /tmp/bla.gz
	reverse loadreplay /tmp/bla.gz
on a machine with an FM-PAC extension. Inconsistently created replays from
before this patch still cannot be loaded, but at least we won't create any
more such replays.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11248 7512d703-071c-0410-b5c2-d2bdd895b7d1