Fix naming of the 'MODE' modifier key
This is part of riihimaki juha's (aka jr) patch for Maemo 5 support.

For all modifier we have a name for the modifier key(s) itself and for the
modifier in combination with other keys. For example LSHIFT, RSHIFT and SHIFT+A, or
LALT, RALT and ALT+A. But for the 'MODE' modifier we only had a name for the key
itself, not in combination with other keys.

This patch renames 'MODE' to 'RMODE' (on azerty keyboards the Alt+Gr key is the
mode key and it's located right from the spacebar). This frees up the name 'MODE'
to indicate mode plus some other key.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11258 7512d703-071c-0410-b5c2-d2bdd895b7d1