Allow typing with non-shift modifiers in the console
This is part of riihimaki juha's (aka jr) patch for Maemo 5 support.

On a N900 the digits 1-9 are typed while holding the 'MODE' modifier key. When
I switch my keyboard to 'azerty' layout, I need to hold the right ALT key to be
able to type characters like |[]. Before this patch, these characters would
appear in the MSX instead of in the console.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11259 7512d703-071c-0410-b5c2-d2bdd895b7d1