Fix bug [2949869] per-machine settings act weird on new machine creation
When a new setting is created, the initial value is taken from the saved values
in settings.xml (if the value is not in this file, we take the default value
from the setting's constructor). Internally in the openMSX code, this file is
represented by the SettingsConfig class.

Before this patch only when a setting is destroyed or right before a
'save_settings' command is executed, we synchronize the setting's value with the
SettingsConfig object. But this way SettingsConfig is not always in sync with
the latest value of the Settings and thus it triggers the problems as described
in the bug report. So the solution is simple (as always once you understand the
problem): do a sync on every value change, that way it's always in sync.

The whole settings stuff has grown over time into a reasonably complex
structure. I'm sure it can be simplified quite a bit. But close before a
release is not a good time to do such big refactorings.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@11260 7512d703-071c-0410-b5c2-d2bdd895b7d1